% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.d.R
\name{cvm.d}
\alias{cvm.d}
\title{Cramer-von Mises test}
\usage{
cvm.d(x, y)
}
\arguments{
\item{x}{one sample}

\item{y}{one sample}
}
\value{
a numeric of Cramer-von Mises statistic\code{n}
}
\description{
compute the Cramer-von Mises statistic for two samples
}
\examples{
\dontrun{
attach(cars)
cvm.d(speed,dist)
detach(cars)
}
}
