% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.dRay.R
\name{MC.dRay}
\alias{MC.dRay}
\title{Rayleigh distribution function}
\usage{
MC.dRay(x, sigma = 1, R = 10000, antithetic = TRUE)
}
\arguments{
\item{x}{vector of quantiles}

\item{sigma}{scale parameter of Rayleigh distribution}

\item{R}{size of Monte Carlo method}

\item{antithetic}{antithetic, if TRUE, use antithetic method}
}
\value{
vector of Rayleigh distribution function\code{n}
}
\description{
Compute the Rayleigh distribution function by Monte Carlo method
}
\examples{
\dontrun{
MC1<-numeric(1000)
MC2<-numeric(1000)
for(i in 1:1000) {
MC1[i]<-MC.dRay(1,antithetic = FALSE)
MC2[i]<-MC.dRay(1)
}
print(c(sd(MC1),sd(MC2)))
}
}
