% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{Table}
\usage{
table2(x, y)
}
\arguments{
\item{x}{one object}

\item{y}{one object}
}
\value{
a table \code{n}
}
\description{
build a contingency table of the counts at each combination of factor levels with two objects
}
\examples{
\dontrun{
x<-c(1,2,2,3,4)
y<-c(1,2,3,3,4)
table2(x,y)
}
}
